check file name

+++ app/Http/Controllers/Home/FileManagerController.php
@@ -23,6 +23,9 @@ class FileManagerController extends Controller
             case "pass":
                 $prompt = "upload successful! please choose another action";
                 break;
+
+            case "badname":
+                $prompt = "bad file name. must be fastq format.";
         }
         return view("FileManager", compact("files", "prompt"));
     }
@@ -33,10 +36,12 @@ class FileManagerController extends Controller
         $file = $request->file("filename");
         $category = $request->input('fileCategory')[0];
         $newFileName = $request->input('newFileName');
+        $file_name_len = strlen($newFileName);
+        if (substr_compare($newFileName, '.fastq', $file_name_len-6) != 0){
+            return redirect("FileManager/badname");
+        }
         $file->storePubliclyAs($category, $newFileName, ['disk' => 'uploads']); // todo: save large files
         $output = exec('../../t1g5/bin/python3 ../scripts/filecheck.py ../storage/app/uploads/'.$category.'/'.$newFileName);
-        echo("<script>console.log('".'../../t1g5/bin/python3 ../scripts/filecheck.py ../storage/app/uploads/'.$category.'/'.$newFileName."');</script>");
-        echo("<script>console.log('".$output."');</script>");
         if ($output == "fail") {
             Storage::delete($category . "/" . $newFileName);
             return redirect("FileManager/fail");
